fix(models): 金额字段 Float → Numeric(12,2)，消除浮点精度风险
- integration.py: POSTransaction/SupplierOrder/MemberSync/ReservationSync 共 9 个金额字段
- marketing_campaign.py: budget/actual_cost/revenue_generated
- ai_model.py: price/price_paid/revenue_share
- decision_log.py: cost_impact/revenue_impact
- store.py: monthly_revenue_target

保留 Float 的非金额字段: latitude/longitude/area/accuracy/confidence/ratio 等

+++ apps/api-gateway/src/models/ai_model.py
@@ -1,7 +1,7 @@
 """
 AI 模型市场数据模型
 """
-from sqlalchemy import Column, String, Integer, Float, DateTime, Text, Enum as SQLEnum, ForeignKey
+from sqlalchemy import Column, String, Integer, Float, DateTime, Text, Enum as SQLEnum, ForeignKey, Numeric
 from sqlalchemy.orm import relationship
 from datetime import datetime
 import enum
@@ -45,7 +45,7 @@ class AIModel(Base):
     model_level = Column(SQLEnum(ModelLevel), nullable=False, index=True)
     industry_category = Column(String(50), index=True)
     description = Column(Text)
-    price = Column(Float, default=0.0, comment="年费（元）")
+    price = Column(Numeric(12, 2), default=0.0, comment="年费（元）")
     training_stores_count = Column(Integer, default=0)
     training_data_points = Column(Integer, default=0)
     accuracy = Column(Float, default=0.0, comment="模型准确率（%）")
@@ -66,7 +66,7 @@ class ModelPurchaseRecord(Base):
     model_id = Column(String(36), ForeignKey("ai_models.id"), nullable=False, index=True)
     purchase_date = Column(DateTime, default=datetime.utcnow, nullable=False)
     expiry_date = Column(DateTime, nullable=False)
-    price_paid = Column(Float, default=0.0)
+    price_paid = Column(Numeric(12, 2), default=0.0)
     status = Column(SQLEnum(PurchaseStatus), default=PurchaseStatus.ACTIVE, index=True)
 
     store = relationship("Store", backref="model_purchases")
@@ -83,7 +83,7 @@ class DataContributionRecord(Base):
     data_points_contributed = Column(Integer, default=0)
     quality_score = Column(Float, default=0.0, comment="数据质量评分（0-100）")
     contribution_date = Column(DateTime, default=datetime.utcnow, nullable=False)
-    revenue_share = Column(Float, default=0.0, comment="分成金额（元）")
+    revenue_share = Column(Numeric(12, 2), default=0.0, comment="分成金额（元）")
 
     store = relationship("Store", backref="data_contributions")
     model = relationship("AIModel", back_populates="contributions")
